# backend/docs/swagger.yaml
definitions:
...
  internal_handlers.SubmitAnswersRequest:
    properties:
      answers:
        additionalProperties: true
        type: object
    required:
    - answers
    type: object
...
info:
  contact: {}
paths:
...
  /auth/onboarding/response:
    get:
      description: Возвращает сохраненные ответы пользователя на опросник
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить ответы пользователя
      tags:
      - onboarding
  /auth/onboarding/submit:
    post:
      consumes:
      - application/json
      description: Обрабатывает ответы пользователя, создает заведение и настраивает
        систему
      parameters:
      - description: Ответы на вопросы опросника
        in: body
        name: answers
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.SubmitAnswersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Отправить ответы опросника
      tags:
      - onboarding
...
  /orders:
    get:
      description: Возвращает список заказов
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      - description: Статус заказа
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить список заказов
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Создает новый заказ
      parameters:
      - description: Данные заказа
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Создать заказ
      tags:
      - orders
  /orders/{id}:
    get:
      description: Возвращает заказ по ID
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Получить заказ по ID
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Обновляет данные заказа
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Обновить заказ
      tags:
      - orders
  /orders/{id}/pay:
    post:
      description: Оплачивает заказ
      parameters:
      - description: ID заказа
        in: path
        name: id
        required: true
        type: string
      - description: Данные оплаты
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Оплатить заказ
      tags:
      - orders
  /statistics/abc-analysis:
    get:
      description: Возвращает ABC анализ товаров
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить ABC анализ
      tags:
      - statistics
  /statistics/products:
    get:
      description: Возвращает статистику по товарам
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить статистику по товарам
      tags:
      - statistics
  /statistics/sales:
    get:
      description: Возвращает статистику продаж
      parameters:
      - description: Начальная дата
        in: query
        name: start_date
        type: string
      - description: Конечная дата
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Получить статистику продаж
      tags:
      - statistics
...
swagger: "2.0"