# backend/internal/handlers/order_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"go.uber.org/zap"

	"github.com/yourusername/arc/backend/internal/usecases"
)

type OrderHandler struct {
	usecase *usecases.OrderUseCase
	logger  *zap.Logger
}

func NewOrderHandler(usecase *usecases.OrderUseCase, logger *zap.Logger) *OrderHandler {
	return &OrderHandler{
		usecase: usecase,
		logger:  logger,
	}
}

// List возвращает список заказов
// @Summary Получить список заказов
// @Description Возвращает список заказов
// @Tags orders
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Param status query string false "Статус заказа"
// @Success 200 {object} map[string]interface{}
// @Router /orders [get]
func (h *OrderHandler) List(c *gin.Context) {
	// TODO: Implement list orders logic
	c.JSON(http.StatusOK, gin.H{"data": []interface{}{}})
}

// Get возвращает заказ по ID
// @Summary Получить заказ по ID
// @Description Возвращает заказ по ID
// @Tags orders
// @Produce json
// @Security Bearer
// @Param id path string true "ID заказа"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Router /orders/{id} [get]
func (h *OrderHandler) Get(c *gin.Context) {
	id, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}

	// TODO: Implement get order logic
	c.JSON(http.StatusOK, gin.H{"id": id})
}

// Create создает новый заказ
// @Summary Создать заказ
// @Description Создает новый заказ
// @Tags orders
// @Accept json
// @Produce json
// @Security Bearer
// @Param request body object true "Данные заказа"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /orders [post]
func (h *OrderHandler) Create(c *gin.Context) {
	// TODO: Implement create order logic
	c.JSON(http.StatusCreated, gin.H{"message": "order created"})
}

// Update обновляет заказ
// @Summary Обновить заказ
// @Description Обновляет данные заказа
// @Tags orders
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path string true "ID заказа"
// @Param request body object true "Данные для обновления"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /orders/{id} [put]
func (h *OrderHandler) Update(c *gin.Context) {
	// TODO: Implement update order logic
	c.JSON(http.StatusOK, gin.H{"message": "order updated"})
}

// Pay оплачивает заказ
// @Summary Оплатить заказ
// @Description Оплачивает заказ
// @Tags orders
// @Produce json
// @Security Bearer
// @Param id path string true "ID заказа"
// @Param request body object false "Данные оплаты"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /orders/{id}/pay [post]
func (h *OrderHandler) Pay(c *gin.Context) {
	// TODO: Implement pay order logic
	c.JSON(http.StatusOK, gin.H{"message": "order paid"})
}