# backend/internal/handlers/statistics_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/yourusername/arc/backend/internal/usecases"
)

type StatisticsHandler struct {
	usecase *usecases.StatisticsUseCase
	logger  *zap.Logger
}

func NewStatisticsHandler(usecase *usecases.StatisticsUseCase, logger *zap.Logger) *StatisticsHandler {
	return &StatisticsHandler{
		usecase: usecase,
		logger:  logger,
	}
}

// GetSales возвращает статистику продаж
// @Summary Получить статистику продаж
// @Description Возвращает статистику продаж
// @Tags statistics
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /statistics/sales [get]
func (h *StatisticsHandler) GetSales(c *gin.Context) {
	// TODO: Implement sales statistics logic
	c.JSON(http.StatusOK, gin.H{"data": map[string]interface{}{}})
}

// GetProducts возвращает статистику по товарам
// @Summary Получить статистику по товарам
// @Description Возвращает статистику по товарам
// @Tags statistics
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /statistics/products [get]
func (h *StatisticsHandler) GetProducts(c *gin.Context) {
	// TODO: Implement products statistics logic
	c.JSON(http.StatusOK, gin.H{"data": map[string]interface{}{}})
}

// GetABCAnalysis возвращает ABC анализ
// @Summary Получить ABC анализ
// @Description Возвращает ABC анализ товаров
// @Tags statistics
// @Produce json
// @Security Bearer
// @Param start_date query string false "Начальная дата"
// @Param end_date query string false "Конечная дата"
// @Success 200 {object} map[string]interface{}
// @Router /statistics/abc-analysis [get]
func (h *StatisticsHandler) GetABCAnalysis(c *gin.Context) {
	// TODO: Implement ABC analysis logic
	c.JSON(http.StatusOK, gin.H{"data": map[string]interface{}{}})
}